Universal term equality operator
Operators with arbitrary argument sorts don't fit into a standard
signature, which is why term equality must be implemented as a special
case.

The contract of lookup-op must be changed because sort constraints can
now be #f in the input arity (for universal variables, to be implemented
later) and in the output rank (for the rank of ==).