Bug fix in context construction
In the original version, all included contexts were merged, and then
the new declarations were added to this merged version. Therefore the
terms inside the rules and equations of the merged includes had
the signature of the merged includes rather than the signature of the
context under construction. This would lead to error messages
of the type "term not allowed in context" during rewriting.