Less strict checks on transformations
Transformations are now allowed to introduce variables.  At application
time, these variables are checked for potential conflicts with the vars
in the input term.

Note: conflicts with the context's varset are impossible by construction
because term-local vars are checked for conflicts at term creation
time. This also applies to terms inside equations and transformations.