Fix a weird problem
Sometimes context? would return #f for an argument that was definitely
a context. Apparently this is related to the way how stuff was
provided from module contexts.rkt. The same happened with sort-graph?

Some quick tests suggest that providing the whole structure
(struct-out ...) solves the problem as well, but the data structure
is supposed to remain opaque.