Fix insertion of included contexts
When a context is inserted with transformations, the transformations
must also be applied to included contexts. This is now finally done:
includes are replaced by recursive inserts that apply the same
transformation everywhere.

This change required to remove the hide-vars transformation, which was
replaced by two different insertions, insert-use and insert-extend.